+++ projects/gpn-ui/src/app/core/authorization/authorization.data.ts
@@ -41,6 +41,6 @@ export class AuthorizationData {
   }
 
   public getRobotState(): Observable<{ state: boolean }> {
-    return this.http.get<{ state: boolean }>(`${api}/audit/robot`);
+    return this.http.get<{ state: boolean }>(`${api}/account/robot`);
   }
 }
