+++ projects/gpn-ui/src/app/features/pre-audit/list-pre-audit/list-pre-audit.component.html
@@ -22,7 +22,9 @@
 
       <ng-container matColumnDef="checkType">
         <th mat-header-cell *matHeaderCellDef > Вид проверки </th>
-        <td mat-cell *matCellDef="let audit"> {{audit.checkType | translate}} </td>
+        <td mat-cell *matCellDef="let audit"> <ng-container *ngFor="let checkType of audit.checkType; let isLast=last">
+          {{checkType | translate}}{{isLast ? '' : ', '}}
+        </ng-container> </td>
       </ng-container>
 
       <ng-container matColumnDef="files">
