+++ projects/gpn-ui/src/app/features/pre-audit/list-pre-audit/list-pre-audit.component.ts
@@ -145,5 +145,6 @@ export class ListPreAuditComponent implements OnInit {
       name: 'checkType',
       value: this.checked
     });
+    this.loadAuditsPage();
   }
 }
