+++ projects/server/controller/account-controller.js
@@ -4,3 +4,7 @@ exports.getUserInfo = async (req, res) => {
   await logger.log(req, res, 'Вход в приложение');
   res.send(res.locals.user);
 };
+
+exports.getRobotState = (req, res) => {
+  res.send({ state: global.robot });
+};
