+++ projects/server/controller/audit-controller.js
@@ -5,7 +5,6 @@ const logger = require('../core/logger');
 const { Audit, Document, Subsidiary, Risk } = require('../models');
 const parser = require('../services/parser-service');
 const translations = require('../../gpn-ui/src/assets/i18n/ru.json');
-const roboService = require('../services/robo-service');
 
 exports.postAudit = async (req, res) => {
   let audit = new Audit(req.body);
@@ -248,59 +247,6 @@ exports.fetchAudits = async (req, res) => {
           a.secondaryStatus.date = robot.new_deadline;
         }
       }
-      return a;
-    });
-
-    res.send({ count, audits });
-  } catch (err) {
-    logger.logError(req, res, err, 500);
-  }
-};
-
-exports.fetchPreAudits = async (req, res) => {
-  try {
-    let sort;
-    if (req.query.column) {
-      sort = {
-        [req.query.column]: req.query.sort === 'asc' ? 1 : -1
-      };
-    }
-    const where = [];
-
-    const checkType = req.query.checkType;
-
-    where.push({ 'pre-check': { $exists: true } });
-
-    let count;
-    let audits;
-    count = await Audit.countDocuments({ $and: where });
-    audits = await Audit.find({ $and: where })
-      .lean()
-      .setOptions({
-        skip: +req.query.skip,
-        limit: +req.query.take,
-        sort
-      });
-
-    audits = audits.map(a => {
-      a.authorName = a.author && a.author.name;
-      const roles = a.author && a.author.roles;
-      const role = roles.forEach(
-        role => role._id === 4 || role._id === 5 || role._id === 6
-      );
-      a.checkType = [];
-      switch (role._id) {
-        case 4:
-          a.checkType.push('InsiderControl');
-          break;
-        case 5:
-          a.checkType.push('InterestControl');
-          break;
-        case 6:
-          a.checkType.push('ProjectControl');
-          break;
-      }
-      a.files = 'Файлики';
       return a;
     });
 
@@ -825,16 +771,6 @@ exports.postSelectedViolations = async (req, res) => {
   }
 };
 
-exports.uploadFiles = async (req, res) => {
-  try {
-    const author = res.locals.user;
-    await roboService.postFiles(req.body, author);
-    res.status(201).json();
-  } catch (err) {
-    logger.logError(req, res, err, 500);
-  }
-};
-
 getViolations = selectedRows => {
   let violations = selectedRows;
 
@@ -944,8 +880,4 @@ getViolations = selectedRows => {
     violationModel.push(violation);
   });
   return violationModel;
-};
-
-exports.getRobotState = (req, res) => {
-  res.send({ state: global.robot });
 };
