+++ projects/server/routers/account-router.js
@@ -2,5 +2,6 @@ const router = require('express').Router();
 const controller = require('../controller/account-controller');
 
 router.get('/user', controller.getUserInfo);
+router.get('/robot', controller.getRobotState);
 
 module.exports = router;
