+++ projects/server/routers/audit-router.js
@@ -16,8 +16,5 @@ router.post('/exportConclusion', controller.exportConclusion);
 router.get('/conclusion', controller.getConclusion);
 router.put('/conclusion', controller.postConclusion);
 router.get('/fetch', controller.fetchAudits);
-router.get('/fetchPreAudits', controller.fetchPreAudits);
 router.put('/selectedViolations', controller.postSelectedViolations);
-router.get('/robot', controller.getRobotState);
-router.post('/uploadFiles', controller.uploadFiles);
 module.exports = router;
