+++ projects/server/routers/preAudit-router.js
@@ -1,0 +1,6 @@
+const router = require('express').Router();
+const controller = require('../controller/preAudit-controller');
+
+router.get('/fetchPreAudits', controller.fetchPreAudits);
+router.post('/uploadFiles', controller.uploadFiles);
+module.exports = router;
