+++ projects/server/server.js
@@ -25,6 +25,7 @@ const auditRouter = require('./routers/audit-router');
 const documentRouter = require('./routers/document-router');
 const eventRouter = require('./routers/event-router');
 const handBookRouter = require('./routers/handBook-router');
+const preAuditRouter = require('./routers/preAudit-router');
 
 const CONTEXT = `/${process.env.CONTEXT || 'gpn-ui'}`;
 
@@ -49,6 +50,7 @@ app.use('/api/audit', auditRouter);
 app.use('/api/document', documentRouter);
 app.use('/api/event', eventRouter);
 app.use('/api/handbook', handBookRouter);
+app.use('/api/preAudit', preAuditRouter);
 
 const listen = async err => {
   if (err) return console.log(err);
